==> 01_EMA_idio_processing.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r Github}
# usethis::create_from_github("https://github.com/Sfglatt/ED_idiographic_networks.git",
# destdir = /Github/ED_idiographic_net")
```

```{r Workspace and packages}
# clear work space
# rm(list=ls())

# set the participant ID name and file path for all forthcoming files 
id_name = "PR042" 

file_path <- file.path("Data", id_name)
dir.create(file_path)

if (!require("cowplot")) {install.packages("cowplot"); require("cowplot")} 
if (!require("devtools")) {install.packages("devtools"); require("devtools")} 
# devtools::install_github("cecusack/clairecleans") 
if (!require("devtools")) {install.packages("devtools"); require("devtools")} 
if (!require("ggpubr")) {install.packages("ggpubr"); require("ggpubr")} #
if (!require("graphicalVAR")) {install.packages("graphicalVAR"); require("graphicalVAR")} 
if (!require("grid")) {install.packages("grid"); require("grid")} # figs
if (!require("gridExtra")) {install.packages("gridExtra"); require("gridExtra")} 
if (!require("here")) {install.packages("here"); require("here")}
if (!require("imputeTS")) {install.packages("imputeTS"); require("imputeTS")} 
if (!require("psych")) {install.packages("psych"); require("psych")} 
if (!require("tidyverse")) {install.packages("tidyverse"); require("tidyverse")} 
if (!require("tseries")) {install.packages("tseries"); require("tseries")} 
if (!require("qgraph")) {install.packages("qgraph"); require("qgraph")}
```

```{r packages}

# function to lag dat by 1
# lagpad <- function(x, k) {
#      c(rep(NA, k), x)[1 : length(x)] 
#  }

# Import the data
morn_import <- read.csv(dir(pattern='.csv$')[1], na.strings = c("", " "))
dim(morn_import) # 31 x 75
aft_import <- read.csv(dir(pattern='.csv$')[2], na.strings = c("", " "))
dim(aft_import)  # 62 x 73
eve_import <- read.csv(dir(pattern='.csv$')[3], na.strings = c("", " "))
dim(eve_import)  # 30 x 97

morn <- morn_import
aft <- aft_import
eve <- eve_import

# Clean up the data
# Morning EMA
if (any(names(morn) == "Duration..seconds..from.scheduled.to.completion.time")) {
  morn <- morn[-c(which(colnames(morn) == "Device.ID"):which(colnames(morn) == "Issued.Time"),
                  which(colnames(morn) == "Duration..seconds..from.scheduled.to.completion.time"):which(colnames(morn) == "Location"))]
} else {
  morn <- morn[-c(which(colnames(morn) == "Device.ID"):which(colnames(morn) == "Issued.Time"),
                  which(colnames(morn) == "Duration..minutes."):which(colnames(morn) == "Location"))]
}

# Afternoon EMA
if (any(names(aft) == "Duration..seconds..from.scheduled.to.completion.time")) {
  aft <- aft[-c(which(colnames(aft) == "Device.ID"):which(colnames(aft) == "Issued.Time"),
                which(colnames(aft) == "Duration..seconds..from.scheduled.to.completion.time"):which(colnames(aft) == "Location"))]
} else {
  aft <- aft[-c(which(colnames(aft) == "Device.ID"):which(colnames(aft) == "Issued.Time"),
                which(colnames(aft) == "Duration..minutes."):which(colnames(aft) == "Location"))]
}

# Evening EMA
if (any(names(eve) == "Duration..seconds..from.scheduled.to.completion.time")) {
  eve <- eve[-c(which(colnames(eve) == "Device.ID"):which(colnames(eve) == "Issued.Time"),
                which(colnames(eve) == "Duration..seconds..from.scheduled.to.completion.time"):which(colnames(eve) == "Location"))]
} else {
  eve <- eve[-c(which(colnames(eve) == "Device.ID"):which(colnames(eve) == "Issued.Time"),
                which(colnames(eve) == "Duration..minutes."):which(colnames(eve) == "Location"))]
}

# setdiff(names(morn), names(aft)) # vas 2 slept well. current menst
# setdiff(names(morn), names(eve)) # sleep and menst
# setdiff(names(eve), names(morn))

# Read in a file with morning EMA
Morn_name <- scan("Morn_name.txt", what = "character", sep = "\n")

if (dim(morn)[2] == 72) {
  colnames(morn) <- Morn_name[-70]
} else if (dim(morn)[2] == 71) {
  colnames(morn) <- Morn_name[-c(68, 69)]
} else {
  colnames(morn) <- Morn_name[-c(68:70)]
}

colnames(aft) <- colnames(morn)[-c(3:4)]

# Read in a file with evening EMA
Eve_name <- scan("Eve_name.txt", what = "character", sep = "\n")
if (dim(morn)[2] == 72) {
  colnames(eve)[1:70] <- Morn_name[-c(3:4, 70)]
  colnames(eve)[71:ncol(eve)] <- Eve_name
} else if (dim(morn)[2] == 71) {
  colnames(eve)[1:69] <- Morn_name[-c(3:4, 68:69)]
  colnames(eve)[70:ncol(eve)] <- Eve_name
} else {
  colnames(eve)[1:68] <- Morn_name[-c(3:4, 68:70)]
  colnames(eve)[69:ncol(eve)] <- Eve_name
}

# Merge the EMA timepoints
dat <- plyr::rbind.fill(morn, aft, eve)
# & Arrange un order
dat <- dat %>% dplyr::arrange(ethica_time)

# Add a day and beep variable
dat$day <- dat$ethica_time
dat <- dat %>% relocate(day, .after = ethica_time)
# Just remove completely empty rows
if (dim(dat[is.na(dat$ethica_time), ])[1] > 0) {
  dat <- dat[!is.na(dat$ethica_time), ]
} else {
  dat <- dat
}

if (sub(".* ([A-Z])", "\\1", dat$ethica_time)[1] == "CDT" | sub(".* ([A-Z])", "\\1", dat$ethica_time)[1] == "CST") {
  dat$ethica_time <- as.POSIXct(dat$ethica_time, tz = "America/Chicago")
} else if (sub(".* ([A-Z])", "\\1", dat$ethica_time)[1] == "EST" | sub(".* ([A-Z])", "\\1", dat$ethica_time)[1] == "EDT") {
  dat$ethica_time <- as.POSIXct(dat$ethica_time, tz = "America/New_York")
} else if (sub(".* ([A-Z])", "\\1", dat$ethica_time)[1] == "MST" | sub(".* ([A-Z])", "\\1", dat$ethica_time)[1] == "MDT") {
  dat$ethica_time <- as.POSIXct(dat$ethica_time, tz = "America/Denver")
} else {
  dat$ethica_time <- as.POSIXct(dat$ethica_time, tz = "America/Los_Angeles")
}

dat$ethica_time_utc <- dat$ethica_time
dat <- dat %>% relocate(ethica_time_utc, .after = ethica_time)
dat$ethica_time_utc <- format(dat$ethica_time_utc, tz = "UTC")
dat$ethica_time_utc <- as.POSIXct(format(dat$ethica_time_utc, tz = "UTC"), tz = "UTC")
attributes(dat$ethica_time)$tzone # est
attributes(dat$ethica_time_utc)$tzone # UTC

dat$day <- gsub("\\s*\\w*$", "", dat$day)
dat <- dat %>% separate(day, c("date", "time"), sep = " ", fill = "right")

# Duplicate sleep response 
dat <- dat %>%
  dplyr::group_by(date) %>%
  fill(sleep, .direction = "down") %>%
  fill(menst_yn, .direction = "down") %>% 
  dplyr::ungroup()

# dat[dat=="(ID 1) Yes"]<- "yes"
# dat[dat=="(ID 2) No"]<- "no"

dat$ID = id_name

sum(is.na(dat[,-c(1:5)]))/prod(dim(dat[,-c(1:5)])) 

# Only keep continuous 0-100 items; remove y/n vars and the ones at the end of the day
dat <- dat[-c(which(colnames(dat) == "location"):ncol(dat))]
dat <- dat[, -grep("yn$|describe$", colnames(dat))] 

# Check missingness
sum(is.na(dat[,-c(1:5)]))/prod(dim(dat[,-c(1:5)])) 

# look
head(dat, 12) # Look at 3 days worth of data (x4 daily assessments)
```

```{r Day and beep variables}
# Convert date column to class date
dat$date <- as.Date(dat$date, "%Y-%m-%d")

# Day variable
dat <- dat %>% 
  mutate(date = lubridate::as_date(date)) %>% 
  mutate(dayvar = cumsum(!duplicated(date))) %>% 
  mutate(beepconsec = seq(1:n())) %>% 
  relocate(dayvar, .after = time) %>% 
  relocate(beepconsec, .after = dayvar)

# Beep variable
dat <- dat %>% 
  group_by(date) %>%
  mutate(beepvar = seq(1:n())) %>% 
  ungroup() %>% 
  relocate(beepvar, .after = dayvar)

describe(dat[6:8]) 
```

```{r Deal with time}
# Dealing with time:

# Make a new variable that places time in the row below (a lag)
dat <- dat %>% mutate(lag = lag(ethica_time)) %>% relocate(lag, .after = ethica_time)

# Calculate time differences
dat$tdif <- as.numeric(difftime(strptime(dat$ethica_time, "%Y-%m-%d %H:%M:%S"), strptime(dat$lag, "%Y-%m-%d %H:%M:%S"))) # subtracting time from time lag

# Check. Lag should have the values of start moved down one row
head(dat[, c(which(colnames(dat) == "ethica_time"), which(colnames(dat) == "lag"), which(colnames(dat) == "tdif"))]) 

# Remove night lag
dat$tdif <- ifelse(dat$beepvar == 1, dat$tdif == 0, dat$tdif)
dat <- dat %>% relocate(tdif, .after = lag)

# Replace NA for first col
dat$tdif[is.na(dat$tdif)] <- 0

dat %>% select(ethica_time, lag, tdif)

# Calculate cumulative sum of numeric elapsed time 
dat$cumsumT <- cumsum(dat$tdif)
dat <- dat %>% relocate(cumsumT, .after = tdif)
# Look at it
dat[, c(1:10)]
dat <- dat[-c(which(colnames(dat) == "date"), which(colnames(dat) == "time"))]

# Deal with sleep!!
dat <- dat %>% 
  dplyr::mutate(sleep = 100 - sleep)

# Save data in CSV form
filetype <- ".csv"
name <- paste0(id_name, "_rawwithtime")
filename <- paste(name, Sys.Date(), filetype, sep = '')
full_filename <- file.path(file_path, filename)
write.csv(dat, file = full_filename, row.names = FALSE)
```

```{r Impute data}
# `na_kalman` is better imputation for time-series data 
# cc  Mansueto et al. (2022). Investigating the Feasibility of Idiographic Networks. Psychological Methods;  https://psyarxiv.com/hgcz6/ 10.1037/met0000466

# import dataset (if already ran the above processing chunks): 
# PR003_dat <- read.csv("PR008_rawwithtime2024-05-26.csv")

dat_2 <- dat

# impute


# Call na_kalman additionally supplying the specified list as optim.control variable
# new_param <- list(maxit = 3000)
imp_kalman <- na_kalman(dat_2[10:ncol(dat_2)], model = "StructTS", smooth = TRUE,#optim.control = new_param
                        )
imp_kalman <- cbind(dat_2[1:9],imp_kalman) 

dat_2 <- imp_kalman 

# Check missingness
sum(is.na(dat_2[-c(1:9)]))/prod(dim(dat_2[-c(1:9)])) # 0 is good
# if it isn't 0, find out where the missing values are
sum(is.na(dat_2[-c(1:9)])) 
dat_2[rowSums(is.na(dat_2[-c(1:9)])) > 0, ] # dayvar = 1, beepvar =1
colSums(is.na(dat_2))

#### save imputed data ####
name <- paste0(id_name, "_imputed_rawwithtime")
filename <- paste(name, Sys.Date(), filetype, sep = '')
full_filename <- file.path(file_path, filename)
write.csv(dat_2, file = full_filename, row.names = FALSE)
```
