==> 05d_data.Rmd <==
---
title: "05d_data_processing"
author: "Sglatt"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r packages}
if (!require("dplyr")) {install.packages("dplyr"); require("dplyr")} 
if (!require("lubridate")) {install.packages("lubridate"); require("lubridate")}
if (!require("purrr")) {install.packages("purrr"); require("purrr")}
if (!require("tidyr")) {install.packages("tidyr"); require("tidyr")} 
```

# EMA data
```{r EMA}
setwd("Raw_data\\05d_raw_data\\Raw_EMA")

# import the data
morn_import <- read.csv(dir(pattern='.csv$')[1], na.strings = c("", " "))
dim(morn_import) 
aft_import <- read.csv(dir(pattern='.csv$')[2], na.strings = c("", " "))
dim(aft_import)  
eve_import <- read.csv(dir(pattern='.csv$')[3], na.strings = c("", " "))
dim(eve_import)  

morn <- morn_import
aft <- aft_import
eve <- eve_import

View(morn)

# Remove extra (empty) metadata columns
morn <- morn[, !grepl(".Metadata.", colnames(morn))]
aft <- aft[, !grepl(".Metadata.", colnames(aft))]
eve <- eve[, !grepl(".Metadata.", colnames(eve))]

View(morn)
View(aft)
View(eve)

# remove other extra columns
morn <- morn[, !colnames(morn) %in% c(
  "Unanswered.Status", 
  "Device.ID", "Device.Manufacturer", "Device.Model", 
  "Device.Last.Used", "Device.App.Version", "Device.App.Update.Date",
  "UUID", "Activity.Version", "Prompt.Time", "Expiry.Time", "Status", 
  "Triggering.Logic.ID", "Triggering.Logic.Type", 
  "Duration..seconds..from.scheduled.to.completion.time", 
  "Duration..seconds..from.first.response.to.completion.time",
  "Location", 
  "X.83_MAQ_1..Indoors", "X.83_MAQ_2..Outdoors"
)]

aft <- aft[, !colnames(aft) %in% c(
  "Unanswered.Status", 
  "Device.ID", "Device.Manufacturer", "Device.Model", 
  "Device.Last.Used", "Device.App.Version", "Device.App.Update.Date",
  "UUID", "Activity.Version", "Prompt.Time", "Expiry.Time", "Status", 
  "Triggering.Logic.ID", "Triggering.Logic.Type", 
  "Duration..seconds..from.scheduled.to.completion.time", 
  "Duration..seconds..from.first.response.to.completion.time",
  "Location", 
  "X.83_MAQ_1..Indoors", "X.83_MAQ_2..Outdoors"
)]

eve <- eve[, !colnames(eve) %in% c(
  "Unanswered.Status", 
  "Device.ID", "Device.Manufacturer", "Device.Model", 
  "Device.Last.Used", "Device.App.Version", "Device.App.Update.Date",
  "UUID", "Activity.Version", "Prompt.Time", "Expiry.Time", "Status", 
  "Triggering.Logic.ID", "Triggering.Logic.Type", 
  "Duration..seconds..from.scheduled.to.completion.time", 
  "Duration..seconds..from.first.response.to.completion.time",
  "Location"
)]

# remove csv item tags for merging
colnames(morn) <- gsub("\\.6758$", "", colnames(morn))
colnames(morn)

colnames(aft) <- gsub("\\.6963$", "", colnames(aft))
colnames(aft)

colnames(eve) <- gsub("\\.6964$", "", colnames(eve))
colnames(eve)

# check that all datasets have the same columns between them / for discrepancies
setdiff(names(morn), names(aft)) 
setdiff(names(morn), names(eve)) 
setdiff(names(aft), names(eve)) 
```

# Deal with time
```{r EMA time}
# Merge the EMA timepoints
dat <- plyr::rbind.fill(morn, aft, eve)

# & Arrange in order
dat <- dat %>% dplyr::arrange(Session.Scheduled.Time)
# View(dat)
colnames(dat)
View(dat)

# Add a day and beep variable
dat$day <- dat$Session.Scheduled.Time
dat <- dat %>% relocate(day, .after = Session.Scheduled.Time)

# this if {} time call is from Claire
if (sub(".* ([A-Z])", "\\1", dat$Session.Scheduled.Time)[1] == "CDT" | sub(".* ([A-Z])", "\\1", dat$Session.Scheduled.Time)[1] == "CST") {
  dat$Session.Scheduled.Time <- as.POSIXct(dat$Session.Scheduled.Time, tz = "America/Chicago")
} else if (sub(".* ([A-Z])", "\\1", dat$Session.Scheduled.Time)[1] == "EST" | sub(".* ([A-Z])", "\\1", dat$Session.Scheduled.Time)[1] == "EDT") {
  dat$Session.Scheduled.Time <- as.POSIXct(dat$Session.Scheduled.Time, tz = "America/New_York")
} else if (sub(".* ([A-Z])", "\\1", dat$Session.Scheduled.Time)[1] == "MST" | sub(".* ([A-Z])", "\\1", dat$Session.Scheduled.Time)[1] == "MDT") {
  dat$Session.Scheduled.Time <- as.POSIXct(dat$Session.Scheduled.Time, tz = "America/Denver")
} else {
  dat$Session.Scheduled.Time <- as.POSIXct(dat$Session.Scheduled.Time, tz = "America/Los_Angeles")
}

# day/beep vars
dat$day <- gsub("\\s*\\w*$", "", dat$day)
dat <- dat %>% separate(day, c("date", "time"), sep = " ", fill = "right")

dat$date <- as.Date(dat$date, "%Y-%m-%d")

dat <- dat %>%
  mutate(date = as_date(date)) %>%  #
  group_by(Participant.Label) %>%  
  mutate(
    dayvar = cumsum(!duplicated(date)),  
    beepconsec = seq(1, n()) 
  ) %>%
  relocate(dayvar, .after = time) %>%  
  relocate(beepconsec, .after = dayvar) %>%  
  
  # Beep variable
  group_by(date) %>%  
  mutate(beepvar = seq(1, n())) %>%  
  ungroup() %>%  
  relocate(beepvar, .after = dayvar)  

View(dat)

# Remove empty EMA (all items starting with "X."?) rows
dat_clean <- dat[!apply(dat[, grepl("^X.", colnames(dat))], 1, function(x) all(is.na(x))), ]
# View(dat_clean)
colnames(dat_clean)
```

# Baseline data/OCI scores
```{r baseline}
baseline <- read.csv("Raw_data/05d_raw_data/R15 baseline.csv")
View(baseline)

baseline <- baseline %>%
  mutate(participant_id = recode(participant_id, 
                                    "pr053" = "PR053", 
                                    "pr182" = "PR182", 
                                    "pr231" = "PR231",
                                    "pr074" = "PR074", 
                                    "Pr248" = "PR248")) # when merging baseline with EMA - found that some ID's were missing since they were labeled incorrectly.


# confirm item ranges 0-4
apply(baseline[, grepl("^oci", colnames(baseline))], 2, range, na.rm = TRUE)

# make total sum column
baseline$oci_sum <- rowSums(as.data.frame(lapply(baseline[, paste0("oci_", 1:18)], function(x)
  as.numeric(gsub("\\s", "", x)))),
  na.rm = TRUE)

summary(baseline$oci_sum) # 0-63

# making a new column with oci > or equal to 21 = 1 and < 21 = 0
baseline$oci_21_or_higher <- ifelse(baseline$oci_sum >= 21, 1, 0)

table(baseline$oci_21_or_higher) # 1 = 60

# filter baseline dataset to OCI >= 21
baseline_oci_filtered <- baseline[baseline$oci_21_or_higher == 1, ]
```

# Match baseline OCI participants/scores to EMA and filter the EMA datset
```{r OCI baseline and EMA}
# rename label/id columns to be consistent in EMA and baseline datasets
baseline <- baseline %>%
  rename(Participant.Label = participant_id)

# add OCI cutoff and total columns to dat (with empty rows removed)
dat_clean <- dat_clean %>%
  left_join(baseline %>%
              select(Participant.Label, oci_21_or_higher, oci_sum), 
            by = "Participant.Label")

# check if anyone is missing OCI
dat_clean %>%
  filter(is.na(oci_sum)) %>%
  summarize(unique_missing_participants = n_distinct(Participant.Label)) %>%
  pull(unique_missing_participants) # 10 EMA IDs don't have OCI in baseline

# look at who they are
(missing_oci_sum_ids <- dat_clean %>%
  filter(is.na(oci_sum)) %>%
  select(Participant.Label) %>%
  distinct())

table(baseline$Participant.Label) # okay, PR053, 182, 231, and 074 are called "pr" or "Pr" in the baseline, so they don't map onto EMA. Go back in the above chunk and correct their IDs...lol

# filter to oci_21_or_higher = 1
dat_clean_oci_filter <- dat_clean %>%
  filter(oci_21_or_higher == 1)

# add OCI cutoff and total column to dat (with empty rows still there)
dat <- dat %>%
  left_join(baseline %>%
              select(Participant.Label, oci_21_or_higher, oci_sum), 
            by = "Participant.Label")

# filter to oci_21_or_higher = 1
dat_oci_filter <- dat %>%
  filter(oci_21_or_higher == 1)

# check how many unique IDs are in the EMA OCI > 20 data 
n_distinct(dat$Participant.Label)
n_distinct(dat_clean_oci_filter$Participant.Label) # 50
```

# save csvs
```{r csvs}
# move ID and OCI scores to the beginning of the datasets
dat <- dat %>%
  select(Participant.Label, oci_21_or_higher, oci_sum, everything())

dat_oci_filter <- dat_oci_filter %>%
  select(Participant.Label, oci_21_or_higher, oci_sum, everything())

dat_clean <- dat_clean %>%
  select(Participant.Label, oci_21_or_higher, oci_sum, everything())

dat_clean_oci_filter <- dat_clean_oci_filter %>%
  select(Participant.Label, oci_21_or_higher, oci_sum, everything())

View(dat_clean_oci_filter)

write.csv(dat, 
          file = paste0("05d_Output/Created_data/EMA_", format(Sys.Date(), "%Y_%m_%d"), ".csv"), 
          row.names = TRUE) # version *with* empty rows (no OCI filter)

write.csv(dat_oci_filter, 
          file = paste0("05d_Output/Created_data/EMA_OCI21_", format(Sys.Date(), "%Y_%m_%d"), ".csv"), 
          row.names = TRUE) # version *with* empty rows and OCI >= 21

write.csv(dat_clean, 
          file = paste0("05d_Output/Created_data/EMA_clean_", format(Sys.Date(), "%Y_%m_%d"), ".csv"), 
          row.names = TRUE) # version *without* empty rows (no OCI filter)

write.csv(dat_clean_oci_filter, 
          file = paste0("05d_Output/Created_data/EMA_OCI21_clean_", format(Sys.Date(), "%Y_%m_%d"), ".csv"), 
          row.names = TRUE) # version *without* empty rows and OCI >= 21

dat_clean_oci_filter %>%
  split(.$Participant.Label) %>%
  walk2(names(.), ~ write.csv(.x, file.path("05d_Output/Created_data/EMA_OCI21_clean/",
                                            paste0("EMA_", .y, ".csv")),
                              row.names = TRUE)) # person-level CSVs *without* empty rows and OCI >= 21
```


```{r analysis dataset}
# put together a version of the dataset EMA_OCI21_2025_03_04 (with missing rows included) that follows the below format with day and beep variables. Please include the following variables (you don't have to rename them): Q5, Q7, Q9, Q10,  Q11, Q13, Q14, Q15, Q21, Q22, Q23, Q25, Q32, Q41, Q42, Q43, Q44, Q45, Q47, Q48, Q49, Q66, Q70
colnames(dat_oci_filter) 

dat_oci_filter <- dat_oci_filter %>%
  group_by(Participant.Label, dayvar) %>%               
  mutate(beepvar_upd = row_number()) %>%  # make beepvar right                
  ungroup()  

dat_oci_filtered_IV <- dat_oci_filter %>%
  select(
    Participant.Label, 
    Session.Scheduled.Time, 
    date, 
    time, 
    dayvar, 
    beepvar_upd,
    matches("^X\\.(5|7|9|10|11|13|14|15|21|22|23|25|32|41|42|43|44|45|47|48|49|66|70)_VAS..Question\\.(5|7|9|10|11|13|14|15|21|22|23|25|32|41|42|43|44|45|47|48|49|66|70)\\.of\\.Survey$")
  ) %>%
  select(
    Participant.Label, 
    Session.Scheduled.Time, 
    date, 
    time, 
    dayvar, 
    beepvar_upd,
    X.5_VAS..Question.5.of.Survey,
    X.7_VAS..Question.7.of.Survey,
    X.9_VAS..Question.9.of.Survey,
    X.10_VAS..Question.10.of.Survey,
    X.11_VAS..Question.11.of.Survey,
    X.13_VAS..Question.13.of.Survey,
    X.14_VAS..Question.14.of.Survey,
    X.15_VAS..Question.15.of.Survey,
    X.21_VAS..Question.21.of.Survey,
    X.22_VAS..Question.22.of.Survey,
    X.23_VAS..Question.23.of.Survey,
    X.25_VAS..Question.25.of.Survey,
    X.32_VAS..Question.32.of.Survey,
    X.41_VAS..Question.41.of.Survey,
    X.42_VAS..Question.42.of.Survey,
    X.43_VAS..Question.43.of.Survey,
    X.44_VAS..Question.44.of.Survey,
    X.45_VAS..Question.45.of.Survey,
    X.47_VAS..Question.47.of.Survey,
    X.48_VAS..Question.48.of.Survey,
    X.49_VAS..Question.49.of.Survey,
    X.66_VAS..Question.66.of.Survey,
    X.70_VAS..Question.70.of.Survey
  )

dat_oci_filtered_IV <- dat_oci_filtered_IV %>%
  arrange(Participant.Label, dayvar, beepvar_upd) 

View(dat_oci_filtered_IV)

dat_oci_filtered_IV <- dat_oci_filtered_IV %>%
  rename(
    ID = Participant.Label,    
    day = dayvar,              
    beep = beepvar_upd      
  )


write.csv(dat_oci_filtered_IV, 
          file = paste0("05d_Output/Created_data/EMA_OCI21_vars_", format(Sys.Date(), "%Y_%m_%d"), ".csv"), 
          row.names = TRUE) 
```
