==> 11_hrv_sui.Rmd <==
---
title: "11_hrv_sui"
author: "Sglatt"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Packages
```{r packages}
if (!require("dplyr")) {
  install.packages("dplyr")
  require("dplyr")
}
if (!require("lme4")) {
  install.packages("lme4")
  require("lme4")
}
if (!require("lmerTest")) {
  install.packages("lmerTest")
  require("lmerTest")
}
if (!require("ggplot2")) {
  install.packages("ggplot2")
  require("ggplot2")
}
if (!require("lubridate")) {
  install.packages("lubridate")
  require("lubridate")
}
if (!require("splines")) {
  install.packages("splines")
  require("splines")
}
```

# Data and folder
```{r data}
dat <- read.csv("11_raw_materials/ema_hrv.csv")

colnames(dat)

if (!dir.exists("11_Output")) {
  dir.create("11_Output")
}
```

# Clean data
```{r clean}
# Rename timestamp column
dat <- dat %>%
  rename(timestamp = Date.and.time)

# Clean participant IDs
dat <- dat %>%
  mutate(
    participant_id = toupper(participant_id),
    participant_id = gsub("[^A-Za-z0-9]", "", participant_id)
  )

# Convert Unix timestamp to POSIXct datasetetime
head(dat$timestamp)

# dat <- dat %>%
#  mutate(
#    timestamp = as.POSIXct(timestamp, origin = "1970-01-01", tz = "UTC"),
#    date_only = as.Date(timestamp)
#  )


dat <- dat %>%
  mutate(
    timestamp = ymd_hms(timestamp, tz = "UTC"),
    date_only = as.Date(timestamp)
  )

head(dat$date_only)

# Sort by participant and datasete
dat <- dat %>%
  arrange(participant_id, date_only)

# Make a unique day ID for each participant
dat <- dat %>%
  group_by(participant_id) %>%
  mutate(dayID = dense_rank(date_only)) %>%
  ungroup()

# Count number of surveys per day per participant
dat <- dat %>%
  group_by(participant_id, date_only) %>%
  mutate(survey = row_number()) %>%
  ungroup()
```

# Variables
```{r vars}
# Passive ideation average from items 1-3
dat <- dat %>%
  mutate(
    passiveideation_avg = rowMeans(cbind(
      passiveideation1..sliderNegPos.,
      passiveideation2..sliderNegPos.,
      passiveideation3..sliderNegPos.
    ), na.rm = TRUE)
  )

# SA binary variable
dat <- dat %>%
  mutate(SA = case_when(
    typesb..multipleChoice. == "No, I haven't done anything to try to kill myself" ~ 1,
    typesb..multipleChoice. == "Started to do something to try to kill myself but someone or something stopped me" ~ 2,
    typesb..multipleChoice. == "Started to do something to try to kill myself but stopped myself" ~ 3,
    typesb..multipleChoice. == "Took steps to prepare to kill myself" ~ 4,
    typesb..multipleChoice. == "Tried to kill myself" ~ 5,
    TRUE ~ NA_real_
  ))

# SA binary - 0 is no endorsement
dat$SA_binary <- ifelse(dat$SA == 1, 0, 1)
```

# Check HRV distribution
```{r hrv}
summary(dat$median_hrv_preceding_30min)
hist(dat$median_hrv_preceding_30min,
  breaks = 30,
  main = "HRV Distribution",
  xlab = "Median HRV preceding 30min"
)

# Scale HRV (mean = 0, SD = 1)
dat$median_hrv_scaled <- scale(dat$median_hrv_preceding_30min)

# Break the HRV variable into between-person and within-person predictors

# This function calculates person-level means. This will create a level-2 variable that can be used in tandem with person-centered means. This is useful if you are interested in both the within-person and between-person effects.
pmean <- function(ID, var) {
  centered <- ave(var, ID, FUN = function(x) mean(x, na.rm = TRUE))
  return(centered)
}

# This function centers on person-means.(How much each individual score differs from the average score for **that individual**)
pcenter <- function(ID, var) {
  centered <- var - ave(var, ID, FUN = function(x) mean(x, na.rm = TRUE))
  return(centered)
}

# Create a between-person and within-person variable for the predictor variable (HRV)

# Participant-mean scores (between-person) for HRV
dat$median_hrv_preceding_30min_pmeans <- pmean(
  dat$participant_id,
  dat$median_hrv_preceding_30min
)

# person-centered mean scores (within-person) for HRV
dat$median_hrv_preceding_30min_pcent <- pcenter(
  dat$participant_id,
  dat$median_hrv_preceding_30min
)

# Participant-mean scores (between-person) for FAD
dat$fearofdeath_pmean <- pmean(
  dat$participant_id,
  dat$fearofdeath..sliderNegPos.
)

# person-centered mean scores (within-person) for FAD
dat$fearofdeath_pcent <- pcenter(
  dat$participant_id,
  dat$fearofdeath..sliderNegPos.
)
```

# Models with passive SI as the outcome
```{r hrv passive}
# Null model
passiveSI_null <- lmer(
  passiveideation_avg ~ 1
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(passiveSI_null)

# HRV predicting future Passive SI
model_PassiveSI <- lmer(
  passiveideation_avg ~ median_hrv_preceding_30min
    + (1 | participant_id),
  data = dat, REML = FALSE
)

summary(model_PassiveSI)

# HRV predicting future Passive SI with FAD condition
model_PassiveSI_condition <- lmer(
  passiveideation_avg ~ median_hrv_preceding_30min + fearofdeath..sliderNegPos.
    + (1 | participant_id),
  data = dat, REML = FALSE
)

summary(model_PassiveSI)

# module comparison Null vs Conditioned
anova(passiveSI_null, model_PassiveSI_condition)


# With scaled HRV
model_PassiveSI_scaled <- lmer(
  passiveideation_avg ~ median_hrv_preceding_30min + (1 | participant_id:dayID),
  data = dat,
  REML = FALSE
)
summary(model_PassiveSI_scaled)

# Add random slopes
model_PassiveSI_randslope <- lmer(
  passiveideation_avg ~ median_hrv_preceding_30min
    + (1 + median_hrv_preceding_30min | participant_id)
    + (1 | participant_id:dayID),
  data = dat,
  REML = FALSE
)
summary(model_PassiveSI_randslope)

# Partition HRV into between/within person predictors
model_passiveSI_between_within <- lmer(
  passiveideation_avg ~ median_hrv_preceding_30min_pmeans
    + median_hrv_preceding_30min_pcent
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_passiveSI_between_within)

# Check a non-linear model
model_PassiveSI_nl <- lmer(
  passiveideation_avg ~ ns(median_hrv_preceding_30min, knots = 25)
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_PassiveSI_nl)
```

# Models with active SI as the outcome
```{r hrv active}
# Active SI Null Model
activeSI_null <- lmer(
  activeideation1..sliderNegPos. ~ 1
    + (1 | participant_id),
  data = dat, REML = FALSE
)
summary(activeSI_null)

# HRV predicting future Active SI
model_ActiveSI <- lmer(
  activeideation1..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 | participant_id),
  data = dat, REML = FALSE
)
summary(model_ActiveSI)

# Active Ideation Model Comparison
anova(activeSI_null, model_ActiveSI)

# HRV predicting future Active SI with FAD condition
model_ActiveSI_condition <- lmer(
  activeideation1..sliderNegPos. ~ median_hrv_preceding_30min + fearofdeath..sliderNegPos.
    + (1 | participant_id),
  data = dat, REML = FALSE
)

summary(model_PassiveSI)

# module comparison Null vs Conditioned
anova(activeSI_null, model_ActiveSI_condition)

# Add random slopes
model_ActiveSI_randslope <- lmer(
  activeideation1..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 + median_hrv_preceding_30min | participant_id)
    + (1 | participant_id:dayID),
  data = dat,
  REML = FALSE
)
summary(model_ActiveSI_randslope)

# Partition HRV into between/within person predictors
model_ActiveSI_between_within <- lmer(
  activeideation1..sliderNegPos. ~ median_hrv_preceding_30min_pmeans
    + median_hrv_preceding_30min_pcent
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_ActiveSI_between_within)


# Check a non-linear model
model_ActiveSI_nl <- lmer(
  activeideation1..sliderNegPos. ~ ns(median_hrv_preceding_30min, knots = 25)
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_ActiveSI_nl)
```

# Models with suicide methods as the outcome
```{r hrv method}
# Method Null Model
Method_null <- lmer(
  method..sliderNegPos. ~ 1
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(Method_null)

# HRV predicting future Method
model_method <- lmer(
  method..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_method)

# Method Model Comparison
anova(Method_null, model_method)


# HRV predicting future Method with FAD condition
model_Method_condition <- lmer(
  method..sliderNegPos. ~ median_hrv_preceding_30min + fearofdeath..sliderNegPos.
    + (1 | participant_id),
  data = dat, REML = FALSE
)

summary(model_Method_condition)

# module comparison Null vs Conditioned
anova(Method_null, model_Method_condition)

# Add random slopes
model_method_randslope <- lmer(
  method..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 + median_hrv_preceding_30min | participant_id)
    + (1 | participant_id:dayID),
  data = dat,
  REML = FALSE
)
summary(model_method_randslope)

# Partition HRV into between/within person predictors
model_method_between_within <- lmer(
  method..sliderNegPos. ~ median_hrv_preceding_30min_pmeans
    + median_hrv_preceding_30min_pcent
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_method_between_within)

# Check a non-linear model
model_method_nl <- lmer(
  method..sliderNegPos. ~ ns(median_hrv_preceding_30min, knots = 25)
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_method_nl)
```

# Models with fearlessness about death as the outcome
```{r hrv fad}
# FAD Null
FAD_null <- lmer(
  fearofdeath..sliderNegPos. ~ 1
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(FAD_null)

# HRV predicting future FAD
Model_FAD <- lmer(
  fearofdeath..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(Model_FAD)

# FAD Model Comparison
anova(FAD_null, Model_FAD)

# Add random slopes
model_fad_randslope <- lmer(
  fearofdeath..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 + median_hrv_preceding_30min | participant_id)
    + (1 | participant_id:dayID),
  data = dat,
  REML = FALSE
)
summary(model_fad_randslope)

# Partition HRV into between/within person predictors
model_fad_between_within <- lmer(
  fearofdeath..sliderNegPos. ~ median_hrv_preceding_30min_pmeans
    + median_hrv_preceding_30min_pcent
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_fad_between_within)

# Check a non-linear model
model_fad_nl <- lmer(
  fearofdeath..sliderNegPos. ~ ns(median_hrv_preceding_30min, df = 3)
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_fad_nl)

# Check if model fit is improved
anova(FAD_null, model_fad_nl)

# plot this non-linear line
predicted_effect <- ggpredict(model_fad_nl, terms = "median_hrv_preceding_30min [all]")

ggplot(predicted_effect, aes(x = x, y = predicted)) +
  geom_line(color = "blue", linewidth = 1) +
  geom_ribbon(aes(ymin = conf.low, ymax = conf.high), alpha = 0.2, fill = "blue") +
  labs(
    title = "Non-linear relationship between HRV and FAD",
    x = "Median HRV",
    y = "Predicted fear of death"
  ) +
  theme_minimal()
```

# Models with NSSI as the outcome
```{r hrv nssi}
# NSSI Null Model
NSSI_null <- lmer(
  nssibx..sliderNegPos. ~ 1
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(NSSI_null)

Model_NSSI <- lmer(
  nssibx..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(Model_NSSI)

# Model comparison
anova(NSSI_null, Model_NSSI)

# Add random slopes
model_nssi_randslope <- lmer(
  nssibx..sliderNegPos. ~ median_hrv_preceding_30min
    + (1 + median_hrv_preceding_30min | participant_id)
    + (1 | participant_id:dayID),
  data = dat,
  REML = FALSE
)
summary(model_nssi_randslope)

# HRV predicting future NSSI with FAD condition
model_NSSI_condition <- lmer(
  nssibx..sliderNegPos. ~ median_hrv_preceding_30min + fearofdeath..sliderNegPos.
    + (1 | participant_id),
  data = dat, REML = FALSE
)

summary(model_NSSI_condition)

# module comparison Null vs Conditioned
anova(NSSI_null, model_NSSI_condition)

# Partition HRV into between/within person predictors
model_nssi_between_within <- lmer(
  nssibx..sliderNegPos. ~ median_hrv_preceding_30min_pmeans
    + median_hrv_preceding_30min_pcent
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_nssi_between_within)

# Check a non-linear model
model_nssi_nl <- lmer(
  nssibx..sliderNegPos. ~ ns(median_hrv_preceding_30min, knots = 25)
    + (1 | participant_id),
  data = dat,
  REML = FALSE
)
summary(model_nssi_nl)
```

# Models with SA (binary) as the outcome
```{r hrv sa}
# For binary outcome models, use glmer()
SA_binary_null <- glmer(
  SA_binary ~ 1 + (1 | participant_id),
  data = dat,
  family = binomial(link = "logit")
)
summary(SA_binary_null)

# HRV as predictor
Model_SA_binary <- glmer(
  SA_binary ~ median_hrv_preceding_30min
    + (1 | participant_id),
  data = dat,
  family = binomial(link = "logit")
)
summary(Model_SA_binary)

# SA Model Comparison
anova(SA_binary_null, Model_SA_binary)

# HRV predicting SA conditioning FAD
SA_binary_condition <- glmer(
  SA_binary ~ median_hrv_preceding_30min + fearofdeath..sliderNegPos. + (1 | participant_id),
  data = dat,
  family = binomial(link = "logit")
)
summary(SA_binary_condition)
# SA Model Comparison Null vs condition
anova(SA_binary_null, SA_binary_condition)

# Add random slopes
model_SAbinary_randslope <- glmer(
  SA_binary ~ median_hrv_preceding_30min
    + (1 + median_hrv_preceding_30min | participant_id)
    + (1 | participant_id:dayID),
  data = dat,
  family = binomial(link = "logit")
)
summary(model_SAbinary_randslope)

# Partition between/within predictors
model_SAbinary_between_within <- glmer(
  SA_binary ~ median_hrv_preceding_30min_pmeans
    + median_hrv_preceding_30min_pcent
    + (1 | participant_id),
  data = dat,
  family = binomial(link = "logit")
)
summary(model_SAbinary_between_within)
```
